chore: Split UpdateChangeLogTask
Split UpdateChangeLogTask to multiple classes.

APM-2537
